third_party/blink/web_tests/external/wpt/html/webappapis/scripting/events/event-handler-attributes-window-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] shadowed pagereveal (window)
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed pageswap (window)
  assert_equals: body should reflect expected (function) function "() => 0" but got (undefined) undefined
[FAIL] shadowed unload (window)
  assert_equals: body should reflect expected (function) function "() => 0" but got (object) null
[FAIL] shadowed pagereveal removal (window)
  assert_equals: body should reflect expected (object) null but got (undefined) undefined
[FAIL] shadowed pageswap removal (window)
  assert_equals: body should reflect expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
